added some debuging logs the secretv2 bridge.
In the secret v2 bridge dal I change the way the query gets the tags the old way did not get the tags I thought this would solve the bug but since the move service function creates a copy
and assigns that copy to the new path this brakes the link in the tag junction. Added some debugging logs to the SecretV2 bridge.

In the SecretV2 bridge DAL, I changed the way the query retrieves tags. The old method didn’t return the expected tags. I thought this would solve the bug, but since the moveService
function creates a copy and assigns that copy to the new path, it breaks the link in the tag junction.